LeetCode/src/test/java/org/redquark/onlinejudges/leetcode/heap/SeatReservationManagerTest.java
package org.redquark.onlinejudges.leetcode.heap;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.redquark.onlinejudges.leetcode.heap.SeatReservationManager.SeatManager;

public class SeatReservationManagerTest {

    private final SeatManager seatManager = new SeatManager(5);

    @Test
    public void test() {
        assertEquals(1, seatManager.reserve());
        assertEquals(2, seatManager.reserve());
        seatManager.unreserve(2);
        assertEquals(2, seatManager.reserve());
        assertEquals(3, seatManager.reserve());
        assertEquals(4, seatManager.reserve());
        assertEquals(5, seatManager.reserve());
        seatManager.unreserve(5);
    }
}